+++ ec/src/hashing/curve_maps/mod.rs
@@ -1,2 +1,3 @@
+pub mod elligator2;
 pub mod swu;
 pub mod wb;