+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - [\#593](https://github.com/arkworks-rs/algebra/pull/593) (`ark-ec`) Change `AffineRepr::xy()` to return owned values.
 
 ### Features
+- [\#xxx](https://github.com/arkworks-rs/algebra/pull/xxx) (`ark-ec`) Add Elligator2 hash-to-curve map.
 
 ### Improvements